+++ modoboa/core/views/admin.py
@@ -15,7 +15,7 @@ from modoboa.core.models import Extension, Log
 from modoboa.core.utils import new_version_available
 from modoboa.lib import events, parameters
 from modoboa.lib.listing import get_sort_order, get_listing_page
-from modoboa.lib.webutils import (
+from modoboa.lib.web_utils import (
     _render_to_string, render_to_json_response
 )
