+++ modoboa/demo/views.py
@@ -3,8 +3,8 @@ import os
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.utils.translation import ugettext as _
-from modoboa.lib.webutils import render_to_json_response
-from modoboa.lib.emailutils import sendmail_simple, sendmail_fromfile
+from modoboa.lib.web_utils import render_to_json_response
+from modoboa.lib.email_utils import sendmail_simple, sendmail_fromfile
 from modoboa.extensions.admin.lib import needs_mailbox
