+++ modoboa/extensions/admin/views/domain.py
@@ -15,7 +15,7 @@ import reversion
 
 from modoboa.core.models import User
 from modoboa.lib import parameters, events
-from modoboa.lib.webutils import (
+from modoboa.lib.web_utils import (
     _render_to_string, render_to_json_response
 )
 from modoboa.lib.exceptions import (
