+++ modoboa/extensions/admin/views/user.py
@@ -1,7 +1,7 @@
 import reversion
 from django.contrib.auth.decorators import login_required
 from django.utils.translation import ugettext as _
-from modoboa.lib.webutils import (
+from modoboa.lib.web_utils import (
     render_to_json_response, _render_to_string
 )
 from modoboa.extensions.admin.lib import needs_mailbox
