+++ modoboa/extensions/amavis/management/commands/amnotify.py
@@ -10,7 +10,7 @@ from django.utils.translation import ugettext as _
 from modoboa.core.management.commands import CloseConnectionMixin
 from modoboa.core.models import User
 from modoboa.lib import parameters
-from modoboa.lib.emailutils import sendmail_simple
+from modoboa.lib.email_utils import sendmail_simple
 from modoboa.extensions.admin.models import Domain
 
 from modoboa.extensions.amavis import Amavis
