+++ modoboa/extensions/amavis/views.py
@@ -16,7 +16,7 @@ from django.contrib.auth.decorators import login_required
 from modoboa.lib import parameters
 from modoboa.lib.exceptions import BadRequest
 from modoboa.lib.paginator import Paginator
-from modoboa.lib.webutils import (
+from modoboa.lib.web_utils import (
     getctx, render_to_json_response, _render_to_string
 )
 from modoboa.extensions.admin.models import Mailbox, Domain
