+++ modoboa/extensions/radicale/views.py
@@ -11,7 +11,7 @@ from django.contrib.auth.decorators import (
 from django.utils.translation import ugettext as _
 
 from modoboa.lib.listing import get_sort_order, get_listing_page
-from modoboa.lib.webutils import (
+from modoboa.lib.web_utils import (
     _render_to_string, render_to_json_response
 )
 from modoboa.lib.exceptions import NotFound, PermDeniedException
