+++ modoboa/extensions/webmail/lib/imaputils.py
@@ -13,7 +13,7 @@ from django.utils.translation import ugettext as _
 from modoboa.lib import parameters, imap_utf7
 from modoboa.lib.exceptions import InternalError
 from modoboa.lib.connections import ConnectionsManager
-from modoboa.lib.webutils import static_url
+from modoboa.lib.web_utils import static_url
 from ..exceptions import ImapError, WebmailInternalError
 from .fetch_parser import parse_fetch_response
